requirements.txt
astroid==3.0.3
black==23.11.0
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.8
coverage==7.3.2
dill==0.4.0
exceptiongroup==1.3.1
flake8==6.1.0
Flask==3.0.0
Flask-Cors==4.0.0
idna==3.11
importlib_metadata==8.7.0
iniconfig==2.1.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
paho-mqtt==2.1.0
pathspec==0.12.1
platformdirs==4.4.0
pluggy==1.6.0
pulsar-client==3.8.0
pycodestyle==2.11.1
pycryptodome==3.20.0
pyflakes==3.1.0
Pygments==2.19.2
pylint==3.0.2
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
python-dotenv==1.0.0
requests==2.31.0
tomli==2.3.0
tomlkit==0.13.3
tuya-connector-python==0.1.2
typing_extensions==4.15.0
urllib3==2.6.0
websocket-client==1.9.0
Werkzeug==3.1.4
zipp==3.23.0
